* Some significant cleanups and restructuring in this commit, as detailed below * Interface changes: there are now derived classes for each interface type * Address changes: lots of cleanup, one class for each address type * Node changes: add derived classes for each type

+++ src/hagglekernel/ConnectivityEthernet.h
@@ -23,7 +23,7 @@
 #include "Interface.h"
 
 // The haggle connectivity UDP port number.
-#define HAGGLE_UDP_CONNECTIVITY_PORT 9696
+#define HAGGLE_UDP_CONNECTIVITY_PORT ((unsigned short)9696)
 
 struct haggle_beacon {
         u_int32_t seqno;
